Move trackStackedSeriesEnabled next to the other settings trackers
lib/settings/graph.ts is the only caller, and it reached up into
visualizations/analytics.ts for it. It now lives in
lib/settings/analytics.ts with the other setting trackers.
